chore: reorganize clients directory
- split base.py into base_client, base_params
- Make a subdirectory for each provider
- Update imports
