+++ python/mirascope/llm/clients/anthropic/message.py
@@ -1,0 +1,9 @@
+"""Anthropic message types."""
+
+from typing import TypeAlias
+
+from anthropic.types import MessageParam
+
+from ...messages import Message
+
+AnthropicMessage: TypeAlias = Message | MessageParam
