+++ python/mirascope/llm/clients/anthropic/params.py
@@ -1,0 +1,7 @@
+"""Anthropic parameters."""
+
+from ..base import BaseParams
+
+
+class AnthropicParams(BaseParams, total=False):
+    """The parameters for the Anthropic LLM model."""
