+++ python/mirascope/llm/clients/anthropic/registered_llms.py
@@ -1,0 +1,6 @@
+"""Anthropic registered LLM models."""
+
+from typing import Literal, TypeAlias
+
+ANTHROPIC_REGISTERED_LLMS: TypeAlias = Literal["anthropic:claude-3-5-sonnet-latest"]
+"""The Anthropic models registered with Mirascope."""
