+++ python/mirascope/llm/clients/google/message.py
@@ -1,0 +1,9 @@
+"""Google message types."""
+
+from typing import TypeAlias
+
+from google.genai.types import ContentDict, FunctionResponse
+
+from ...messages import Message
+
+GoogleMessage: TypeAlias = Message | ContentDict | FunctionResponse
