+++ python/mirascope/llm/clients/google/params.py
@@ -1,0 +1,7 @@
+"""Google parameters."""
+
+from ..base import BaseParams
+
+
+class GoogleParams(BaseParams, total=False):
+    """The parameters for the Google LLM model."""
