+++ python/mirascope/llm/clients/google/registered_llms.py
@@ -1,0 +1,6 @@
+"""Google registered LLM models."""
+
+from typing import Literal, TypeAlias
+
+GOOGLE_REGISTERED_LLMS: TypeAlias = Literal["google:gemini-2.5-flash"]
+"""The Google models registered with Mirascope."""
