+++ python/mirascope/llm/clients/openai/params.py
@@ -1,0 +1,7 @@
+"""OpenAI parameters."""
+
+from ..base import BaseParams
+
+
+class OpenAIParams(BaseParams, total=False):
+    """The parameters for the OpenAI LLM model."""
