+++ python/mirascope/llm/clients/openai/registered_llms.py
@@ -1,0 +1,6 @@
+"""OpenAI registered LLM models."""
+
+from typing import Literal, TypeAlias
+
+OPENAI_REGISTERED_LLMS: TypeAlias = Literal["openai:gpt-4o-mini"]
+"""The OpenAI models registered with Mirascope."""
